rename BoundingBoxes module to match class name (#7910)
