+++ torchvision/tv_tensors/__init__.py
@@ -1,6 +1,6 @@
 import torch
 
-from ._bounding_box import BoundingBoxes, BoundingBoxFormat
+from ._bounding_boxes import BoundingBoxes, BoundingBoxFormat
 from ._image import Image
 from ._mask import Mask
 from ._torch_function_helpers import set_return_type
